pyproject.toml
[tool.poetry]
name = "darkgpt"
version = "0.1.0"
description = ""
authors = ["Esben Kran <esben@kran.ai>"]
readme = "README.md"
include = ["darkgpt/prompts/**/*.py"]

[tool.poetry.dependencies]
python = "^3.12"

anthropic = "*"
click = "*"
dominate = "*"
google-generativeai = "*"
jinja2 = "*"
matplotlib = "*"
openai = "*"
python-dotenv = "*"
replicate = "*"
ruamel-yaml = "*"
tqdm = "*"

[tool.poetry.group.dev.dependencies]
debugpy = "*"
dvc = { version = "*", extras = ["s3", "ssh"] }
isort = "*"

[tool.poetry.group.system]
optional = true

[tool.poetry.group.system.dependencies]
poetry = "1.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
